(squashme) continue reorganization of graph package
