+++ operator/pkg/gateway-api/helpers/namespaces.go
@@ -71,11 +71,6 @@ func isNamespaceSelected(selector *metav1.LabelSelector, routeNamespace string, 
 	return selectorMatcher.Matches(labels.Set(labelsForNamespace))
 }
 
-// AllowedRouteNamespaces resolves a listener's allowedRoutes.namespaces policy
-// into the set of namespaces whose Routes may attach. A nil result means all
-// namespaces are allowed (From: All); an empty set means none are (From: None).
-// It mirrors IsListenerNamespaceAllowed so pre-resolving the set is equivalent
-// to the per-Route check.
 func AllowedRouteNamespaces(
 	listener gatewayv1.Listener, listenerNamespace string, namespaces []*corev1.Namespace,
 ) map[string]struct{} {
@@ -105,8 +100,6 @@ func AllowedRouteNamespaces(
 	return map[string]struct{}{}
 }
 
-// namespacesMatchingSelector returns the names of namespaces whose labels match
-// the selector. An invalid selector matches nothing.
 func namespacesMatchingSelector(selector *metav1.LabelSelector, namespaces []*corev1.Namespace) map[string]struct{} {
 	allowed := make(map[string]struct{})
 	sel, err := metav1.LabelSelectorAsSelector(selector)
